requirements.txt
asgiref==3.2.7
autopep8==1.4.4
beautifulsoup4==4.9.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.7
django-allauth==0.39.1
django-autoslug==1.9.7
django-bootstrap-form==3.4
django-bootstrap4==1.1.1
django-cors-headers==3.3.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-forms-bootstrap==3.1.0
django-grappelli==2.14.2
django-rest-framework==0.1.0
django-semantic-ui==1.2.3
django-sitetree==1.14.0
django-six==1.0.4
django-static-fontawesome==5.12.1.4
djangorestframework==3.9.1
idna==2.8
mysqlclient==1.4.6
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.2.0
pycodestyle==2.5.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.2.4
stripe==2.27.0
urllib3==1.24.2
